third_party/blink/web_tests/external/wpt/css/css-transitions/Document-getAnimations.tentative-expected.txt

This is a testharness.js-based test.
PASS getAnimations for non-animated content
PASS getAnimations for CSS Transitions
PASS CSS Transitions targetting (pseudo-)elements should have correct order after sorting
FAIL CSS Transitions targetting (pseudo-)elements should have correct order after sorting (::marker) assert_equals: CSS transition on both pseudo-elements and elements are returned expected 5 but got 4
PASS Transitions are not returned after they have finished
Harness: the test ran to completion.
